+++ packages/laconia-acceptance-test/src/UuidIdGenerator.js
@@ -1,4 +1,4 @@
-const uuidv4 = require("uuid/v4");
+const { v4: uuidv4 } = require("uuid");
 
 module.exports = class UuidIdGenerator {
   generate() {